Create two cases
The first case is that when a user wants to delete a transaction and when he presses the delete button, the transaction will be tracked and deleted by the ID of that transaction.
The second case is that when user submits the form, the text and amount will be added as a new transaction and will be placed at top.