Update IncomeExpense.js
Imported neccessary things.
Imported transactions from GlobalContext using useContext,
Created amounts const that have all the amounts,
Created income const and it have the values that are more than 0 and also they all are added,
Created expense const and it have the values that are less than 0 and also they all are added,
Printed income and expense and used .toFixed() in them.